[addons] move AddonProps part to new files
With them become more easy to review the part.
